% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GetElements.R
\name{GetElements}
\alias{GetElements}
\title{Get Elements Defined within Data Warehouse}
\usage{
GetElements(reportsuite.id)
}
\arguments{
\item{reportsuite.id}{Report suite id}
}
\value{
Data frame
}
\description{
Get a data frame of Data Warehouse elements for the specified report suite.
}
\details{
This function requires having a character vector with one valid Report Suite specified.
}
\examples{
\dontrun{
elements <- GetElements("your_report_suite")

}
}

