% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GetMetrics.R
\name{GetMetrics}
\alias{GetMetrics}
\title{Get Metrics Defined within Data Warehouse}
\usage{
GetMetrics(reportsuite.id)
}
\arguments{
\item{reportsuite.id}{Report suite id}
}
\value{
Data frame
}
\description{
Get a data frame of Data Warehouse metrics for the specified report suite.
}
\details{
This function requires having a character vector with one valid Report Suite specified.
}
\examples{
\dontrun{
metrics <- GetMetrics("your_report_suite")

}
}

