% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/IsEnabled.R
\name{IsEnabled}
\alias{IsEnabled}
\title{stuff}
\usage{
IsEnabled(reportsuite.ids)
}
\arguments{
\item{reportsuite.ids}{Report suite id}
}
\value{
Data frame
}
\description{
stuff
}
\details{
stuff
}
\examples{
\dontrun{



}
}

