% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SCAuth.R
\name{SCAuth}
\alias{SCAuth}
\title{Store Credentials for the Adobe Analytics API}
\usage{
SCAuth(key, secret, debug.mode = FALSE)
}
\arguments{
\item{key}{client id from your app in the Adobe Marketing cloud Dev Center OR if you are using auth.method='legacy', then this is the API username (username:company)}

\item{secret}{secret from your app in the Adobe Marketing cloud Dev Center OR if you are using auth.method='legacy', then this is the API shared secret}

\item{debug.mode}{set global debug mode}
}
\value{
Global credentials list 'SC.Credentials'
}
\description{
SCAuth
}
\details{
Authorise and store credentials for the Adobe Analytics API
}
\examples{
\dontrun{
#Legacy authentication
SCAuth("key", "secret")

}
}

