% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/IsEnabled.R
\name{IsEnabled}
\alias{IsEnabled}
\title{Determine Whether DataWarehouse Is Enabled For a User Account/Report Suite}
\usage{
IsEnabled(reportsuite.ids)
}
\arguments{
\item{reportsuite.ids}{Report suite id}
}
\value{
Boolean value
}
\description{
Determine whether data warehouse is enabled for a user account/report suite combination.
}
\details{
Based on testing, it appears that the Adobe DataWarehouse API v1.3 might not be accurate with respect to this function.
Testing has shown a value of 'FALSE' while still being able to access the API. Users are encouraged to try GetDataWarehouseDisplay() from the RSiteCatalyst package if this function returns 'FALSE'.
}
\examples{
\dontrun{

dw_enabled <- IsEnabled("reportsuite")


}
}

